requirements.txt
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
Django==2.0.7
django-filter==2.2.0
djangorestframework==3.10
docker==4.0.2
idna==2.8
public==2019.4.13
pytz==2019.2
requests==2.22.0
self==2019.4.13
six==1.12.0
urllib3==1.25.6
websocket-client==0.56.0
waitress==1.3.1
django-heroku==0.3.1